feat(ch-back): add ClickHouse back-end support

+++ tests/test_schema_cross_backend.py
@@ -102,21 +102,21 @@ def main(local=False):
 
     client = FlightSQLClient(host=host, port=8815, insecure=True)
 
-    # MySQL expected schemas (int64 for integers)
+    # MySQL expected schemas (int32 for INT columns)
     mysql_schemas = {
         'campaigns': {
             'campaign_name': pa.string(),
-            'campaign_id': pa.int64(),
-            'campaign_length': pa.int64(),
+            'campaign_id': pa.int32(),
+            'campaign_length': pa.int32(),
             'budget': pa.float32()
         },
         'impressions': {
-            'impression_id': pa.int64(),
-            'campaign_id': pa.int64(),
-            'clicks': pa.int64(),
-            'days_since_start': pa.int64(),
+            'impression_id': pa.int32(),
+            'campaign_id': pa.int32(),
+            'clicks': pa.int32(),
+            'days_since_start': pa.int32(),
             'revenue': pa.float32(),
-            'conversions': pa.int64()
+            'conversions': pa.int32()
         }
     }
